Migrated to CMake. Started work on FreeBSD port
