+++ src/line.c
@@ -16,6 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <errno.h>
 #include "line.h"
 #include "log.h"
 #include "sys.h"
